address review feedback

+++ component/core/src/main/java/org/exoplatform/social/core/jpa/storage/dao/ActivityDAO.java
@@ -423,7 +423,7 @@ public interface ActivityDAO extends GenericDAO<ActivityEntity, Long> {
    * @param activityId the Id of activity
    * @return the total number of comments
    */
-  long getNumberOfAllComments(long activityId);
+  int getNumberOfAllComments(long activityId);
 
   /**
    *
